+++ vfr-green-app/app/src/main/java/com/blueskycharts/app/map/view/SubMapView.kt
@@ -15,7 +15,7 @@ interface SubMapView {
      * Returns the age of any data beign displayed and resets the age counter prior to rendering
      */
     fun resetRequestedInformationAgeRecord();
-    fun getRequestedInformationAgeSeconds(): Int?;
+    fun getRequestedInformationAgeSeconds(): Long?;
 
     /**
      * Draws or queues the drawing of this sub map.  The number of pixels to be drawn are the region width * height * scale
